backend/requirements.txt
alembic==1.4.3
aniso8601==9.0.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
gunicorn==20.0.4
httpie==2.4.0
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.16.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
neotime==1.7.4
prompt-toolkit==2.0.10
psycopg2==2.8.6
Pygments==2.9.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
rdflib==5.0.0
rdflib-sqlalchemy==0.4.0
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
SQLAlchemy==1.3.20
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==1.0.0
